requirements.txt
APScheduler==3.6.3
asgiref==3.8.1
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.2
django-phonenumber-field==8.0.0
idna==3.10
phonenumbers==8.13.50
pillow==10.3.0
python-telegram-bot==13.15
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
tornado==6.1
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
